refactor(quick-filter): split customize and input modes into separate components
Break QuickFilter's two faces out of one file: QuickFilterInputs (the
normal-mode value strip) and QuickFilterCustomize (the edit-mode field
chips), with QuickFilter as a thin dispatcher owning the shared models.

Customize mode now matches the sibling controls' affordances:
- "Add Filter" reuses the Filter control's #trigger-slot ghost button
- field chips drag-to-reorder via vuedraggable (a real flex-wrap container,
  not a display:contents wrapper, so Sortable measures drop positions
  correctly across wrapped rows; sorting scoped to .quick-filter-chip)
- a Save button lives in QuickFilterCustomize and emits `save` to exit

The host hides its whole right-side cluster (Filter/Sort/Columns + the
customize trigger) while customizing, leaving the strip a focused,
full-width editing surface that Save exits.
